[UPDATE] mularized image upload as a middleware
